pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wintest"
version = "0.1.0"
description = "AI-powered Windows UI testing tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.50.0",
    "Pillow>=12.0.0",
    "PyAutoGUI>=0.9.54",
    "timm>=1.0.0",
    "bitsandbytes>=0.49.0",
    "PyYAML>=6.0",
    "Jinja2>=3.1.0",
    "click>=8.1.0",
    "colorama>=0.4.6",
]

[project.scripts]
wintest = "wintest.ui.cli:cli"

[tool.setuptools.packages.find]
include = ["wintest*"]
